clean aws keyspaces client
